# docs/openapi.yaml
openapi: 3.1.0
info:
  title: A2A Agent Platform API
  version: 0.1.0
  description: |
    Multi-tenant management and external A2A gateway API. Management endpoints use
    a console access token. Gateway endpoints use a tenant API Key. Protocol payloads
    follow A2A Protocol v1.0 and are intentionally represented as open JSON schemas here.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose gateway
security: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200": { $ref: "#/components/responses/AuthSession" }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/register:
    post:
      tags: [Authentication]
      operationId: selfRegister
      description: Creates an unprivileged customer account and immediately starts a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, displayName, password]
              properties:
                email: { type: string, format: email }
                displayName: { type: string, minLength: 2, maxLength: 100 }
                password: { type: string, format: password, minLength: 12 }
      responses:
        "201": { $ref: "#/components/responses/AuthSession" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      description: Rotates the HttpOnly refresh cookie and returns a short-lived access token.
      responses:
        "200": { $ref: "#/components/responses/AuthSession" }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      security: [{ ConsoleToken: [] }]
      responses:
        "204": { description: Session revoked }
  /api/auth/me:
    get:
      tags: [Authentication]
      operationId: currentUser
      security: [{ ConsoleToken: [] }]
      responses:
        "200":
          description: User profile and tenant memberships
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tenants:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/Tenant" },
                    }
  /api/auth/invitations/{token}:
    get:
      tags: [Authentication]
      operationId: inspectInvitation
      parameters: [{ $ref: "#/components/parameters/InvitationToken" }]
      responses:
        "200": { description: Valid invitation metadata }
        "410": { $ref: "#/components/responses/Error" }
  /api/auth/invitations/{token}/activate:
    post:
      tags: [Authentication]
      operationId: activateInvitation
      parameters: [{ $ref: "#/components/parameters/InvitationToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, password]
              properties:
                displayName: { type: string, minLength: 1 }
                password: { type: string, format: password, minLength: 12 }
      responses:
        "200": { $ref: "#/components/responses/AuthSession" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/auth/oidc/exchange:
    post:
      tags: [Authentication]
      operationId: exchangeOidcCode
      responses:
        {
          "200": { $ref: "#/components/responses/AuthSession" },
          "401": { $ref: "#/components/responses/Error" },
        }
...
components:
  securitySchemes:
    ConsoleToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    InvitationToken:
      {
        name: token,
        in: path,
        required: true,
        schema: { type: string, minLength: 16 },
      }
...
  responses:
    AuthSession:
      description: Authenticated session; refresh token is an HttpOnly cookie
      headers:
        Set-Cookie: { schema: { type: string } }
      content:
        application/json:
          schema:
            type: object
            properties:
              accessToken: { type: string }
              accessTokenExpiresIn: { type: integer }
              refreshTokenExpiresAt: { type: string, format: date-time }
              user: { $ref: "#/components/schemas/User" }
    Error:
      description: Structured platform error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    User:
      type: object
      required: [id, email, displayName, status, emailVerified]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        platformRole: { type: string, enum: [platform_admin] }
        status: { type: string, enum: [active, disabled] }
        emailVerified: { type: boolean }
    TenantRole:
      type: string
      enum: [tenant_admin, developer, viewer]
    Tenant:
      type: object
      required: [id, slug, displayName, status]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        displayName: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, suspended] }
        role: { $ref: "#/components/schemas/TenantRole" }
        minuteRequestLimit: { type: integer }
        dailyRequestLimit: { type: integer }
        monthlyRequestLimit: { type: integer }
        concurrentRequestLimit: { type: integer }
        dataRetentionDays: { type: integer, minimum: 7, maximum: 3650 }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
        requestId: { type: string }